# docs/swagger.yaml
...
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  pager.PageModel:
    properties:
      count:
        type: integer
      list: {}
    type: object
...
  schema.AddTagReq:
    properties:
      display_name:
        description: display_name
        maxLength: 35
        type: string
      original_text:
        description: original text
        maxLength: 65536
        type: string
      slug_name:
        description: slug_name
        maxLength: 35
        type: string
      tagg:
        description: tag group
        maxLength: 20
        type: string
    required:
    - display_name
    - original_text
    - slug_name
    - tagg
    type: object
...
  schema.GetFollowingTagsResp:
    properties:
      display_name:
        description: display name
        type: string
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        description: slug name
        type: string
      tag_id:
        description: tag id
        type: string
    type: object
...
  schema.GetTagBasicResp:
    properties:
      display_name:
        type: string
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        type: string
      tag_id:
        type: string
    type: object
...
  schema.GetTagPageResp:
    properties:
      created_at:
        description: created time
        type: integer
      description:
        description: description
        type: string
      display_name:
        description: display_name
        type: string
      excerpt:
        description: excerpt
        type: string
      follow_count:
        description: follower amount
        type: integer
      is_follower:
        description: is follower
        type: boolean
      original_text:
        description: original text
        type: string
      parsed_text:
        description: parsed_text
        type: string
      question_count:
        description: question amount
        type: integer
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        description: slug_name
        type: string
      tag_group:
        description: tag group
        type: string
      tag_id:
        description: tag_id
        type: string
      updated_at:
        description: updated time
        type: integer
    type: object
  schema.GetTagResp:
    properties:
      created_at:
        type: integer
      description:
        type: string
      display_name:
        type: string
      excerpt:
        type: string
      follow_count:
        type: integer
      is_follower:
        type: boolean
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      member_actions:
        items:
          $ref: '#/definitions/schema.PermissionMemberAction'
        type: array
      original_text:
        type: string
      parsed_text:
        type: string
      question_count:
        type: integer
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        type: string
      status:
        type: string
      tag_group:
        description: tag group, exposed with the same name used by the tag list API
        type: string
      tag_id:
        type: string
      tagg:
        type: string
      updated_at:
        type: integer
    type: object
  schema.GetTagSynonymsResp:
    properties:
      member_actions:
        description: MemberActions
        items:
          $ref: '#/definitions/schema.PermissionMemberAction'
        type: array
      synonyms:
        description: synonyms
        items:
          $ref: '#/definitions/schema.TagSynonym'
        type: array
    type: object
...
  schema.PermissionMemberAction:
    properties:
      action:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  schema.RecoverTagReq:
    properties:
      tag_id:
        type: string
    required:
    - tag_id
    type: object
...
  schema.RemoveTagReq:
    properties:
      tag_id:
        description: tag_id
        type: string
    required:
    - tag_id
    type: object
...
  schema.TagItem:
    properties:
      display_name:
        description: display_name
        maxLength: 35
        type: string
      original_text:
        description: original text
        type: string
      slug_name:
        description: slug_name
        maxLength: 35
        type: string
      tag_group:
        description: tag group
        maxLength: 20
        type: string
    required:
    - tag_group
    type: object
...
  schema.TagSynonym:
    properties:
      display_name:
        description: display name
        type: string
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      slug_name:
        description: slug name
        type: string
      tag_id:
        description: tag id
        type: string
    type: object
...
  schema.UpdateTagReq:
    properties:
      display_name:
        description: display_name
        maxLength: 35
        type: string
      edit_summary:
        description: edit summary
        type: string
      original_text:
        description: original text
        type: string
      slug_name:
        description: slug_name
        maxLength: 35
        type: string
      tag_id:
        description: tag_id
        type: string
      tagg:
        description: slug_name
        maxLength: 20
        type: string
    required:
    - tag_id
    type: object
  schema.UpdateTagSynonymReq:
    properties:
      synonym_tag_list:
        description: synonym tag list
        items:
          $ref: '#/definitions/schema.TagItem'
        type: array
      tag_id:
        description: tag_id
        type: string
    required:
    - synonym_tag_list
    - tag_id
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/question/tags:
    get:
      description: get tag list
      parameters:
      - description: tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.GetTagBasicResp'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: get tag list
      tags:
      - Tag
...
  /answer/api/v1/tag:
    delete:
      consumes:
      - application/json
      description: delete tag
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RemoveTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: delete tag
      tags:
      - Tag
    get:
      consumes:
      - application/json
      description: get tag one
      parameters:
      - description: tag id
        in: query
        name: tag_id
        required: true
        type: string
      - description: tag name
        in: query
        name: tag_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetTagResp'
              type: object
      summary: get tag one
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: add tag
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AddTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: add tag
      tags:
      - Tag
    put:
      consumes:
      - application/json
      description: update tag
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update tag
      tags:
      - Tag
  /answer/api/v1/tag/merge:
    post:
      consumes:
      - application/json
      description: merge tag
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AddTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: merge tag
      tags:
      - Tag
  /answer/api/v1/tag/recover:
    post:
      consumes:
      - application/json
      description: recover delete tag
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RecoverTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: recover delete tag
      tags:
      - Tag
  /answer/api/v1/tag/synonym:
    put:
      consumes:
      - application/json
      description: update tag
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateTagSynonymReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update tag
      tags:
      - Tag
  /answer/api/v1/tag/synonyms:
    get:
      description: get tag synonyms
      parameters:
      - description: tag id
        in: query
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetTagSynonymsResp'
              type: object
      summary: get tag synonyms
      tags:
      - Tag
  /answer/api/v1/tags:
    get:
      description: get tags list by slug name
      parameters:
      - collectionFormat: csv
        description: string collection
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.GetTagBasicResp'
                  type: array
              type: object
      summary: get tags list
      tags:
      - Tag
  /answer/api/v1/tags/following:
    get:
      description: get following tag list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.GetFollowingTagsResp'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: get following tag list
      tags:
      - Tag
...
  /answer/api/v1/tags/page:
    get:
      description: get tag page
      parameters:
      - description: page size
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: slug_name
        in: query
        name: slug_name
        type: string
      - collectionFormat: multi
        description: tag groups
        in: query
        items:
          type: string
        name: tag_groups
        type: array
      - description: query condition
        enum:
        - popular
        - name
        - newest
        in: query
        name: query_cond
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetTagPageResp'
                        type: array
                    type: object
              type: object
      summary: get tag page
      tags:
      - Tag
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"